pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "vdi-babysitter"
version = "0.1.0"
description = "CLI tool for automating VDI session management"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "playwright>=1.58",
    "pyyaml>=6.0",
]

[project.scripts]
vdi-babysitter = "vdi_babysitter.main:app"

[tool.setuptools.packages.find]
include = ["vdi_babysitter*"]
